requirements.txt
Flask==2.3.3
flask-cors==4.0.0
azure-identity==1.15.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.4.0